+++ example/lib/ui/ui.dart
@@ -1,3 +1,3 @@
 export 'lang_page.dart';
+export 'nested_example.dart';
 export 'search_page.dart';
-export 'test.dart';