+++ lib/src/implicitly_animated_list.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 
 import 'src.dart';
 
+/// A Flutter ListView that implicitly animates between the changes of two lists.
 class ImplicitlyAnimatedList<E> extends StatelessWidget {
-  /// The data that this [ImplicitlyAnimatedList] should represent.
+  /// The current data that this [ImplicitlyAnimatedList] should represent.
   final List<E> items;
 
   /// Called, as needed, to build list item widgets.
@@ -117,6 +118,8 @@ class ImplicitlyAnimatedList<E> extends StatelessWidget {
   /// The amount of space by which to inset the children.
   final EdgeInsetsGeometry padding;
 
+  /// Creates a Flutter ListView that implicitly animates between the changes
+  /// of two lists.
   const ImplicitlyAnimatedList({
     Key key,
     @required this.items,
@@ -166,7 +169,33 @@ class ImplicitlyAnimatedList<E> extends StatelessWidget {
   }
 }
 
+/// A Flutter Sliver that implicitly animates between the changes of two lists.
 class SliverImplicitlyAnimatedList<E> extends ImplicitlyAnimatedListBase<Widget, E> {
+  /// Creates a Flutter Sliver that implicitly animates between the changes of two lists.
+  /// 
+  /// {@template implicilty_animated_list.constructor}
+  /// The [items] parameter represents the current items that should be displayed in
+  /// the list.
+  ///
+  /// The [itemBuilder] callback is used to build each child as needed. The parent must
+  /// be a [Reorderable] widget.
+  ///
+  /// The [areItemsTheSame] callback is called by the DiffUtil to decide whether two object
+  /// represent the same Item. For example, if your items have unique ids, this method should
+  /// check their id equality.
+  ///
+  /// The [onReorderFinished] callback is called in response to when the dragged item has
+  /// been released and animated to its final destination. Here you should update
+  /// the underlying data in your model/bloc/database etc.
+  ///
+  /// This parameter should not be null.
+  /// The [spawnIsolate] flag indicates whether to spawn a new isolate on which to
+  /// calculate the diff between the lists.
+  ///
+  /// Usually you wont have to specify this value as the MyersDiff implementation will
+  /// use its own metrics to decide, whether a new isolate has to be spawned or not for
+  /// optimal performance.
+  /// {@endtemplate}
   const SliverImplicitlyAnimatedList({
     Key key,
     @required List<E> items,
@@ -203,7 +232,7 @@ class _SliverImplicitlyAnimatedListState<E>
   @override
   Widget build(BuildContext context) {
     return SliverAnimatedList(
-      key: listKey,
+      key: animatedListKey,
       initialItemCount: newData.length,
       itemBuilder: (context, index, animation) {
         final item = dataSet[index];